chore: v2 버전 스위칭을 위한 주석 처리
